+++ src/main/java/mokindang/jubging/project_backend/controller/board/recruitment/RecruitmentBoardController.java
@@ -1,4 +1,4 @@
-package mokindang.jubging.project_backend.controller.board;
+package mokindang.jubging.project_backend.controller.board.recruitment;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
