+++ src/test/java/mokindang/jubging/project_backend/domain/board/vo/StartingDateTest.java
@@ -1,5 +1,6 @@
 package mokindang.jubging.project_backend.domain.board.vo;
 
+import mokindang.jubging.project_backend.domain.board.recruitment.vo.StartingDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
